Split up the compile module
The lifting function has some helpers that don't need to be exposed at all, so
it can be a separate module.  Also move the NodeFilter type into its own module
to reduce the size of the main compile module.
