+++ src/compile.rs
@@ -1,7 +1,9 @@
 mod backend;
 mod errors;
 pub mod interface;
+mod lift;
 mod method_library;
+pub mod node_filter;
 
 use std::collections::HashMap;
 use std::rc::Rc;
@@ -9,49 +11,14 @@ use std::rc::Rc;
 use crate::compile::backend::cpp::CPPTreeInterface;
 use crate::compile::backend::java::JavaTreeInterface;
 use crate::compile::errors::PlanError;
-use crate::compile::interface::{
-    BoundNode, CallableRef, FormalArgument, LanguageType, NodeMatcher, TreeInterface,
-};
+use crate::compile::interface::{BoundNode, CallableRef, NodeMatcher, TreeInterface};
+use crate::compile::lift::transform_node_filter;
 use crate::compile::method_library::{method_impl_for, Handler};
+use crate::compile::node_filter::NodeFilter;
 use crate::query;
-use crate::query::ir::{
-    AggregateOp, CachedRegex, CompOp, Constant, EqualityOp, Expr, Expr_, Repr, Typed,
-};
+use crate::query::ir::{AggregateOp, CompOp, Constant, EqualityOp, Expr, Expr_, Repr, Typed};
 use crate::query::val_type::Type;
 use crate::source_file::{Language, SourceFile};
-
-pub enum NodeFilter {
-    Predicate(NodeMatcher<bool>),
-    PredicateListComputation(NodeMatcher<Vec<bool>>),
-    TypeComputation(NodeMatcher<LanguageType>),
-    TypeListComputation(NodeMatcher<Vec<LanguageType>>),
-    NumericComputation(NodeMatcher<i32>),
-    StringComputation(NodeMatcher<String>),
-    StringListComputation(NodeMatcher<Vec<String>>),
-    RegexComputation(NodeMatcher<CachedRegex>),
-    CallableComputation(NodeMatcher<CallableRef>),
-    ArgumentComputation(NodeMatcher<FormalArgument>),
-    ArgumentListComputation(NodeMatcher<Vec<FormalArgument>>),
-}
-
-impl NodeFilter {
-    /// Return a string representation of the NodeFilter's carried type for debugging purposes
-    pub fn kind(&self) -> String {
-        match self {
-            NodeFilter::Predicate(_) => "bool".into(),
-            NodeFilter::PredicateListComputation(_) => "[bool]".into(),
-            NodeFilter::TypeComputation(_) => "Type".into(),
-            NodeFilter::TypeListComputation(_) => "[Type]".into(),
-            NodeFilter::NumericComputation(_) => "int".into(),
-            NodeFilter::StringComputation(_) => "string".into(),
-            NodeFilter::StringListComputation(_) => "[string]".into(),
-            NodeFilter::RegexComputation(_) => "Regex".into(),
-            NodeFilter::CallableComputation(_) => "Callable".into(),
-            NodeFilter::ArgumentComputation(_) => "Parameter".into(),
-            NodeFilter::ArgumentListComputation(_) => "[Parameter".into(),
-        }
-    }
-}
 
 pub struct Matching;
 
@@ -405,150 +372,6 @@ fn compile_expr(ti: Rc<dyn TreeInterface>, e: &Expr<Typed>) -> anyhow::Result<No
                 }
             }
         }
-    }
-}
-
-fn as_predicate(nf: NodeFilter) -> NodeMatcher<bool> {
-    match nf {
-        NodeFilter::Predicate(nm) => nm,
-        _ => panic!("Impossible, expected Predicate"),
-    }
-}
-
-fn as_string(nf: NodeFilter) -> NodeMatcher<String> {
-    match nf {
-        NodeFilter::StringComputation(nm) => nm,
-        _ => panic!("Impossible, expected String"),
-    }
-}
-
-fn as_type(nf: NodeFilter) -> NodeMatcher<LanguageType> {
-    match nf {
-        NodeFilter::TypeComputation(nm) => nm,
-        _ => panic!("Impossible, expected Type"),
-    }
-}
-
-/// This is the core transformation for each case in `transform_node_filter`
-///
-/// This creates a new `NodeMatcher` that:
-///
-/// 1. Wraps a single element from the base element collection into a suitable
-///    singleton NodeFilter that the transformer can be applied to
-///
-/// 2. Apply the transformer (from the original type to the target type) to
-///    produce a NodeFilter wrapping a single element
-///
-/// 3. Extract the single result at the expected type and append it to the
-///    result collection
-fn transform_body<From, To, W, F, X>(
-    base_elts: &NodeMatcher<Vec<From>>,
-    wrap_one: W,
-    extract_one: F,
-    transformer: Rc<X>,
-) -> NodeMatcher<Vec<To>>
-where
-    From: Clone + 'static,
-    W: Fn(NodeMatcher<From>) -> NodeFilter + 'static,
-    F: Fn(NodeFilter) -> NodeMatcher<To> + 'static,
-    X: Fn(Rc<NodeFilter>) -> anyhow::Result<NodeFilter> + 'static,
-{
-    let xfrm = Rc::clone(&transformer);
-    let args_extract = Rc::clone(&base_elts.extract);
-
-    NodeMatcher {
-        extract: Rc::new(move |ctx, source| {
-            let mut res = Vec::new();
-
-            for arg in args_extract(ctx, source) {
-                // Wrap each element in the carrier type for single elements so
-                // that the scalar processor (the transformer, above) can
-                // process it unmodified
-                let wrapper_matcher = NodeMatcher {
-                    extract: Rc::new(move |_ctx, _source| arg.clone()),
-                };
-
-                let wrapper_filter = wrap_one(wrapper_matcher);
-                let result_filter = xfrm(Rc::new(wrapper_filter)).unwrap();
-                let comp = extract_one(result_filter);
-                let val = (comp.extract)(ctx, source);
-                res.push(val);
-            }
-
-            res
-        }),
-    }
-}
-
-/// A combinator to help lift operations over relations when needed
-///
-/// Most of the operations in the planner work over scalars, but need to be
-/// lifted over some Relational<T> values.  This combinator takes a NodeFilter
-/// that returns a relational value, applies the provided function to each
-/// element in the relation, and re-wraps the results into a relational value of
-/// the appropriate type (specified by the `result_type`)
-///
-/// NOTE: This requires that the result type of `F` is `result_type`. The type
-/// checker ensured this, so we just assume it here.
-///
-/// This function returns None if the given `relation` is not a list.
-/// Otherwise, it always returns a NodeFilter.
-fn transform_node_filter<F>(
-    result_type: Type,
-    relation: &NodeFilter,
-    transformer: Rc<F>,
-) -> Option<NodeFilter>
-where
-    F: Fn(Rc<NodeFilter>) -> anyhow::Result<NodeFilter> + 'static,
-{
-    let xfrm: Rc<F> = Rc::clone(&transformer);
-    match relation {
-        NodeFilter::ArgumentListComputation(arg_list_matcher) => {
-            match result_type.base_if_relational() {
-                Type::PrimString => {
-                    let matcher = transform_body(arg_list_matcher, NodeFilter::ArgumentComputation, as_string, xfrm);
-                    Some(NodeFilter::StringListComputation(matcher))
-                },
-                Type::Type => {
-                    let matcher = transform_body(arg_list_matcher, NodeFilter::ArgumentComputation, as_type, xfrm);
-                    Some(NodeFilter::TypeListComputation(matcher))
-                },
-                _ => panic!("Unsupported conversion from Argument (only String is supported), result type `{}`", result_type)
-            }
-        }
-        NodeFilter::StringListComputation(string_list_matcher) => {
-            match result_type.base_if_relational() {
-                Type::PrimBoolean => {
-                    let matcher = transform_body(
-                        string_list_matcher,
-                        NodeFilter::StringComputation,
-                        as_predicate,
-                        xfrm,
-                    );
-                    Some(NodeFilter::PredicateListComputation(matcher))
-                }
-                _ => {
-                    panic!("Unsupported conversion from `string` to `{}`", result_type);
-                }
-            }
-        }
-        NodeFilter::TypeListComputation(type_list_matcher) => {
-            match result_type.base_if_relational() {
-                Type::PrimString => {
-                    let matcher = transform_body(
-                        type_list_matcher,
-                        NodeFilter::TypeComputation,
-                        as_string,
-                        xfrm,
-                    );
-                    Some(NodeFilter::StringListComputation(matcher))
-                }
-                _ => {
-                    panic!("Unsupported conversion from `Type` to `{}`", result_type);
-                }
-            }
-        }
-        _ => None,
     }
 }
